requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-groq>=0.0.1
langchain-google-genai>=0.0.5
langgraph>=0.0.40

# YouTube and transcript processing
pytube>=15.0.0
youtube-transcript-api>=0.6.0

# Environment and utilities
python-dotenv>=0.19.0
requests>=2.28.0
urllib3>=1.26.0

# Optional: For enhanced features
numpy>=1.21.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
